requirements.txt
langchain==0.1.17
langchain-google-genai==0.0.8
python-dotenv==1.1.0
